+++ tests/test_numerbay.py
@@ -1,4 +1,5 @@
-import os
+import nacl.utils
+from nacl.public import PrivateKey, SealedBox
 import pytest
 import responses
 
@@ -189,7 +190,18 @@ def test_upload_artifact(api, tmpdir):
     api.user_id = 2
     api.token = "Token"
 
+    # mock product search
     product_id = 2
+    data = {
+        'total': 1,
+        'data': [{
+            "id": product_id,
+            "name": "mymodel",
+            "sku": "numerai-predictions-mymodel",
+        }]
+    }
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/products/search-authenticated", json=data)
+
     artifact_id = 3
     data = {"url": "https://uploadurl", "id": artifact_id}
     responses.add(responses.POST, f"{API_ENDPOINT_URL}/products/{product_id}/artifacts/generate-upload-url", json=data)
@@ -223,7 +235,18 @@ def test_upload_artifact_df(api):
     api.user_id = 2
     api.token = "Token"
 
+    # mock product search
     product_id = 2
+    data = {
+        'total': 1,
+        'data': [{
+            "id": product_id,
+            "name": "mymodel",
+            "sku": "numerai-predictions-mymodel",
+        }]
+    }
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/products/search-authenticated", json=data)
+
     artifact_id = 3
     data = {"url": "https://uploadurl", "id": artifact_id}
     responses.add(responses.POST, f"{API_ENDPOINT_URL}/products/{product_id}/artifacts/generate-upload-url", json=data)
@@ -252,24 +275,124 @@ def test_upload_artifact_df(api):
 
 
 @responses.activate
-def test_download_artifact(api, tmpdir):
+def test_upload_encrypted_artifact(api, tmpdir):
+    key_pair = nacl.public.PrivateKey.generate()
+    public_key = key_pair.public_key.encode(encoder=nacl.encoding.Base64Encoder)
+
     api.user_id = 2
     api.token = "Token"
-    api.show_progress_bars = False
 
-    product_id = 4
-    artifact_id = 3
-    data = "https://downloadurl"
-    responses.add(responses.GET, f"{API_ENDPOINT_URL}/products/{product_id}/artifacts/{artifact_id}/generate-download-url", json=data)
-    responses.add(responses.GET, "https://downloadurl")
+    # mock product search
+    product_id = 2
+    data = {
+        'total': 1,
+        'data': [{
+            "id": product_id,
+            "name": "mymodel",
+            "sku": "numerai-predictions-mymodel",
+            "use_encryption": True
+        }]
+    }
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/products/search-authenticated", json=data)
+
+    # mock order search
+    data = {
+        'total': 1,
+        'data': [{
+            "date_order": "2021-12-25T06:34:58.047278",
+            "round_order": 296,
+            "quantity": 1,
+            "price": 9,
+            "currency": "NMR",
+            "mode": "file",
+            "stake_limit": None,
+            "submit_model_id": None,
+            "submit_model_name": None,
+            "submit_state": None,
+            "chain": "ethereum",
+            "from_address": "0x00000000000000000000000000000fromaddress",
+            "to_address": "0x0000000000000000000000000000000toaddress",
+            "transaction_hash": "0x09bd2a0f814a745f62cb35f1a41dd18208fb653210ff677e946747a20e5abcdef",
+            "state": "confirmed",
+            "applied_coupon_id": 1,
+            "coupon": None,
+            "coupon_specs": None,
+            "id": 126,
+            "product": {"id": product_id},
+            "buyer": {"id": 2, "username": "myusername"},
+            "buyer_public_key": public_key.decode('ascii'),
+        }]
+    }
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/orders/search", json=data)
+
+    artifact_id = "abc"
+    data = {"url": "https://uploadurl", "id": artifact_id}
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/artifacts/generate-upload-url", json=data)
+    responses.add(responses.PUT, "https://uploadurl")
+    data = {"id": artifact_id}
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/artifacts/{artifact_id}/validate-upload", json=data)
 
     path = tmpdir.join("somefilepath")
     path.write("content")
 
-    # download file with product_id and artifact_id
-    api.download_artifact(dest_path=tmpdir.join("artifact.csv"), product_id=product_id, artifact_id=artifact_id)
+    # upload file with product_id
+    artifacts = api.upload_artifact(str(path), product_id=product_id)
+    assert artifacts[0]["id"] == artifact_id
 
-    # download file with product_full_name only
+    # upload df with product_id
+    df = pd.DataFrame.from_dict({"id": [], "prediction": []})
+    artifacts = api.upload_artifact(df=df, product_id=product_id)
+    assert artifacts[0]["id"] == artifact_id
+
+    # upload file with product_full_name
+    data = {
+        'total': 1,
+        'data': [{
+            "id": product_id,
+            "name": "mymodel",
+            "sku": "numerai-predictions-mymodel",
+            "use_encryption": True
+        }]
+    }
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/products/search-authenticated", json=data)
+    artifacts = api.upload_artifact(str(path), product_full_name="numerai-predictions-mymodel")
+    assert artifacts[0]["id"] == artifact_id
+
+
+@responses.activate
+def test_download_artifact(api, tmpdir):
+    # mock order search
+    data = {
+        'total': 1,
+        'data': [{
+            "date_order": "2021-12-25T06:34:58.047278",
+            "round_order": 296,
+            "quantity": 1,
+            "price": 9,
+            "currency": "NMR",
+            "mode": "file",
+            "stake_limit": None,
+            "submit_model_id": None,
+            "submit_model_name": None,
+            "submit_state": None,
+            "chain": "ethereum",
+            "from_address": "0x00000000000000000000000000000fromaddress",
+            "to_address": "0x0000000000000000000000000000000toaddress",
+            "transaction_hash": "0x09bd2a0f814a745f62cb35f1a41dd18208fb653210ff677e946747a20e5abcdef",
+            "state": "confirmed",
+            "applied_coupon_id": 1,
+            "coupon": None,
+            "coupon_specs": None,
+            "id": 126,
+            "product": {'id': 4},
+            "buyer": {"id": 2, "username": "myusername"}
+        }]
+    }
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/orders/search", json=data)
+
+    # mock product search
+    product_id = 4
+
     data = {
         'total': 1,
         'data': [{
@@ -279,6 +402,21 @@ def test_download_artifact(api, tmpdir):
         }]
     }
     responses.add(responses.POST, f"{API_ENDPOINT_URL}/products/search-authenticated", json=data)
+
+    api.user_id = 2
+    api.token = "Token"
+    api.show_progress_bars = False
+
+
+    artifact_id = 3
+    data = "https://downloadurl"
+    responses.add(responses.GET, f"{API_ENDPOINT_URL}/products/{product_id}/artifacts/{artifact_id}/generate-download-url", json=data)
+    responses.add(responses.GET, "https://downloadurl")
+
+    # download file with product_id and artifact_id
+    api.download_artifact(dest_path=tmpdir.join("artifact.csv"), product_id=product_id, artifact_id=artifact_id)
+
+    # download file with product_full_name only
     data = {
         'total': 1,
         'data': [{
@@ -292,6 +430,35 @@ def test_download_artifact(api, tmpdir):
 
 @responses.activate
 def test_download_artifact_no_active_artifact(api):
+    # mock order search
+    data = {
+        'total': 1,
+        'data': [{
+            "date_order": "2021-12-25T06:34:58.047278",
+            "round_order": 296,
+            "quantity": 1,
+            "price": 9,
+            "currency": "NMR",
+            "mode": "file",
+            "stake_limit": None,
+            "submit_model_id": None,
+            "submit_model_name": None,
+            "submit_state": None,
+            "chain": "ethereum",
+            "from_address": "0x00000000000000000000000000000fromaddress",
+            "to_address": "0x0000000000000000000000000000000toaddress",
+            "transaction_hash": "0x09bd2a0f814a745f62cb35f1a41dd18208fb653210ff677e946747a20e5abcdef",
+            "state": "confirmed",
+            "applied_coupon_id": 1,
+            "coupon": None,
+            "coupon_specs": None,
+            "id": 126,
+            "product": {'id': 4},
+            "buyer": {"id": 2, "username": "myusername"}
+        }]
+    }
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/orders/search", json=data)
+
     api.user_id = 2
     api.token = "Token"
     api.show_progress_bars = False
@@ -316,4 +483,86 @@ def test_download_artifact_no_active_artifact(api):
 
     with pytest.raises(ValueError) as err:
         api.download_artifact(dest_path="artifact.csv", product_full_name="numerai-predictions-somemodel")
-    assert str(err.value).startswith("Failed to resolve a valid artifact ID")
+    assert str(err.value).startswith("Failed to resolve artifact")
+
+@responses.activate
+def test_download_encrypted_artifact(api, tmpdir):
+    key_pair = nacl.public.PrivateKey.generate()
+    public_key = key_pair.public_key.encode(encoder=nacl.encoding.Base64Encoder)
+    private_key = key_pair.encode(encoder=nacl.encoding.Base64Encoder)
+
+    # mock order search
+    artifact_id = "abc"
+    data = {
+        'total': 1,
+        'data': [{
+            "date_order": "2021-12-25T06:34:58.047278",
+            "round_order": 296,
+            "quantity": 1,
+            "price": 9,
+            "currency": "NMR",
+            "mode": "file",
+            "stake_limit": None,
+            "submit_model_id": None,
+            "submit_model_name": None,
+            "submit_state": None,
+            "chain": "ethereum",
+            "from_address": "0x00000000000000000000000000000fromaddress",
+            "to_address": "0x0000000000000000000000000000000toaddress",
+            "transaction_hash": "0x09bd2a0f814a745f62cb35f1a41dd18208fb653210ff677e946747a20e5abcdef",
+            "state": "confirmed",
+            "applied_coupon_id": 1,
+            "coupon": None,
+            "coupon_specs": None,
+            "id": 126,
+            "product": {"id": 4},
+            "buyer": {"id": 2, "username": "myusername"},
+            "buyer_public_key": public_key.decode('ascii'),
+            "artifacts": [{"id": artifact_id}]
+        }]
+    }
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/orders/search", json=data)
+
+    # mock product search
+    product_id = 4
+
+    data = {
+        'total': 1,
+        'data': [{
+            "id": product_id,
+            "name": "somemodel",
+            "sku": "numerai-predictions-somemodel",
+            "use_encryption": True
+        }]
+    }
+    responses.add(responses.POST, f"{API_ENDPOINT_URL}/products/search-authenticated", json=data)
+
+    api.user_id = 2
+    api.token = "Token"
+    api.show_progress_bars = False
+
+    data = "https://downloadurl"
+    content = b'hello'
+    responses.add(responses.GET, f"{API_ENDPOINT_URL}/artifacts/{artifact_id}/generate-download-url", json=data)
+    responses.add(responses.GET, "https://downloadurl", body=SealedBox(key_pair.public_key).encrypt(content))
+
+    # download file with product_id and artifact_id
+    api.download_artifact(dest_path=tmpdir.join("encrypted_artifact_1.csv"), product_id=product_id, artifact_id=artifact_id,
+                          key_base64=private_key.decode('ascii'))
+
+    with open(tmpdir.join("encrypted_artifact_1.csv"), 'rb') as file:
+        assert file.read() == content
+
+    # download file with product_full_name only
+    data = {
+        'total': 1,
+        'data': [{
+            "id": artifact_id,
+        }]
+    }
+    responses.add(responses.GET, f"{API_ENDPOINT_URL}/products/{product_id}/artifacts", json=data)
+
+    api.download_artifact(dest_path=tmpdir.join("encrypted_artifact_2.csv"), product_full_name="numerai-predictions-somemodel", key_base64=private_key.decode('ascii'))
+
+    with open(tmpdir.join("encrypted_artifact_2.csv"), 'rb') as file:
+        assert file.read() == content